+++ goobernetes/test/sample-projects/db-app-single-db/index.js
@@ -1,0 +1,2 @@
+// Sample app with a single database
+console.log('db-app-single-db running');