+++ goobernetes/test/sample-projects/db-app/index.js
@@ -1,0 +1,2 @@
+// Sample app that uses SQLite databases
+console.log('db-app running');